Fix Initial Parameters
