+++ gui/MainWindow.py
@@ -1,12 +1,12 @@
 
 from PyQt5.QtWidgets import QWidget, QVBoxLayout, QMainWindow, QPushButton, QHBoxLayout
 from PyQt5.uic import loadUi
-from gui.InputParameters import InputParameters
-from gui.SecondParams import SecondParams
+from gui.InitialParameters import InitialParameters
+from gui.SecondaryParameters import SecondaryParameters
 from gui.ThirdParams import ThirdParams
 from gui.CalculateButtons import CalculateButtons
 from gui.OutputWindow import OutputWindow
-
+from gui.BaseView import BaseView
 
 class MainWindow(QMainWindow):
     def __init__(self, parent=None):
@@ -15,13 +15,13 @@ class MainWindow(QMainWindow):
         self.setWindowTitle('Простой расчет формовки и параметров индуктора')
         vbox = QVBoxLayout()
         hbox_params = QHBoxLayout()
-        inParam = InputParameters()
-        secondParam = SecondParams()
-        thirdParam = ThirdParams()
+        inParam = InitialParameters()
+        # secondParam = SecondParams()
+        # thirdParam = ThirdParams()
         # hbox.addStretch(1)
         hbox_params.addWidget(inParam)
-        hbox_params.addWidget(secondParam)
-        hbox_params.addWidget(thirdParam)
+        # hbox_params.addWidget(secondParam)
+        # hbox_params.addWidget(thirdParam)
 
         hbox_calculate = QHBoxLayout()
         calc = CalculateButtons()
@@ -34,3 +34,5 @@ class MainWindow(QMainWindow):
         vbox.addLayout(hbox_params)
         vbox.addLayout(hbox_calculate)
         self.centralWidget().setLayout(vbox)
+
+        self.db_view = BaseView()
