+++ gui/SecondParams.ui
@@ -117,63 +117,53 @@
     </layout>
    </item>
    <item row="3" column="0">
-    <widget class="QLabel" name="label_4">
-     <property name="text">
-      <string>R0</string>
-     </property>
-    </widget>
-   </item>
-   <item row="4" column="0">
     <widget class="QLabel" name="label_5">
      <property name="text">
       <string>Размер межвитковой изоляции</string>
      </property>
     </widget>
    </item>
-   <item row="5" column="0">
+   <item row="3" column="1">
+    <widget class="QLineEdit" name="lineEdit_5"/>
+   </item>
+   <item row="4" column="0">
     <widget class="QLabel" name="label_6">
      <property name="text">
       <string>Толщина основной изоляции индуктора</string>
      </property>
     </widget>
    </item>
-   <item row="6" column="0">
+   <item row="4" column="1">
+    <widget class="QLineEdit" name="lineEdit_6"/>
+   </item>
+   <item row="5" column="0">
     <widget class="QLabel" name="label_7">
      <property name="text">
       <string>Толщина воздушного зазора</string>
      </property>
     </widget>
    </item>
-   <item row="7" column="0">
+   <item row="5" column="1">
+    <widget class="QLineEdit" name="lineEdit_7"/>
+   </item>
+   <item row="6" column="0">
     <widget class="QLabel" name="label_8">
      <property name="text">
       <string>Сопротивление материала индуктора</string>
      </property>
     </widget>
    </item>
-   <item row="8" column="0">
+   <item row="6" column="1">
+    <widget class="QLineEdit" name="lineEdit_8"/>
+   </item>
+   <item row="7" column="0">
     <widget class="QLabel" name="label_9">
      <property name="text">
       <string>Индуктивность токоподводов индуктора</string>
      </property>
     </widget>
    </item>
-   <item row="3" column="1">
-    <widget class="QLineEdit" name="lineEdit_4"/>
-   </item>
-   <item row="4" column="1">
-    <widget class="QLineEdit" name="lineEdit_5"/>
-   </item>
-   <item row="5" column="1">
-    <widget class="QLineEdit" name="lineEdit_6"/>
-   </item>
-   <item row="6" column="1">
-    <widget class="QLineEdit" name="lineEdit_7"/>
-   </item>
    <item row="7" column="1">
-    <widget class="QLineEdit" name="lineEdit_8"/>
-   </item>
-   <item row="8" column="1">
     <widget class="QLineEdit" name="lineEdit_9"/>
    </item>
   </layout>
