+++ gui/SecondaryParameters.py
@@ -1,0 +1,10 @@
+from PyQt5.QtCore import pyqtSlot
+from PyQt5.QtWidgets import QWidget
+from PyQt5.uic import loadUi
+from gui.BaseView import BaseView
+
+class SecondaryParameters(QWidget):
+    def __init__(self, parent=None):
+        QWidget.__init__(self, parent)
+        loadUi('./gui/SecondParams.ui', self)
+        print("2")
